verbeteringen

+++ module_1/deel_4/ice_bugs.py
@@ -4,8 +4,8 @@ def convertToEuroText (amount):
 SMALL_PRICE = 1.25
 MEDIUM_PRICE = 2.10
 #Varibelen moet je niet hergebruiken.
-amount = float(input('Hoeveel ijsjes van {} wil je bestellen? '.format(convertToEuroText(SMALL_PRICE))))
-amount_a = float(input('En hoeveel ijsjes van {} wil je bestellen? '.format(convertToEuroText(MEDIUM_PRICE))))
+amount = int(input('Hoeveel ijsjes van {} wil je bestellen? '.format(convertToEuroText(SMALL_PRICE))))
+amount_a = int(input('En hoeveel ijsjes van {} wil je bestellen? '.format(convertToEuroText(MEDIUM_PRICE))))
 totalPrice = amount * SMALL_PRICE + amount_a * MEDIUM_PRICE
 
 print('Dat is dan {} totaal'.format(convertToEuroText(totalPrice)))
